Moved time.go to the api package
